Checkout.com Magento 2 Module - v1.0.11

+++ Model/Service/CallbackService.php
@@ -10,25 +10,25 @@
 
 namespace CheckoutCom\Magento2\Model\Service;
 
+use DomainException;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\App\ObjectManager;
 use Magento\Sales\Api\OrderRepositoryInterface;
 use Magento\Sales\Model\Order\Payment;
 use Magento\Sales\Model\OrderFactory;
+use Magento\Sales\Model\Service\InvoiceService;
+use Magento\Sales\Model\Order\Invoice;
+use Magento\Sales\Api\InvoiceRepositoryInterface;
+use Magento\Customer\Model\CustomerFactory;
+use Magento\Store\Model\StoreManagerInterface;
+use Magento\Sales\Model\Order\Email\Sender\OrderSender;
 use CheckoutCom\Magento2\Model\Adapter\CallbackEventAdapter;
 use CheckoutCom\Magento2\Model\Adapter\ChargeAmountAdapter;
 use CheckoutCom\Magento2\Model\GatewayResponseHolder;
 use CheckoutCom\Magento2\Model\GatewayResponseTrait;
 use CheckoutCom\Magento2\Model\Method\CallbackMethod;
-use DomainException;
-
 use CheckoutCom\Magento2\Gateway\Config\Config as GatewayConfig;
-use Magento\Sales\Model\Service\InvoiceService;
-use Magento\Sales\Model\Order\Invoice;
-use Magento\Sales\Api\InvoiceRepositoryInterface;
 use CheckoutCom\Magento2\Model\Service\StoreCardService;
-use Magento\Customer\Model\CustomerFactory;
-use Magento\Store\Model\StoreManagerInterface;
 
 class CallbackService {
 
@@ -69,9 +69,25 @@ class CallbackService {
      */
     protected $invoiceRepository;
 
+    /**
+     * @var StoreCardService
+     */
     protected $storeCardService;
+
+    /**
+     * @var CustomerFactory
+     */
     protected $customerFactory;
+
+    /**
+     * @var StoreManagerInterface
+     */
     protected $storeManager;
+
+    /**
+     * @var OrderSender
+     */
+    protected $orderSender;
 
     /**
      * CallbackService constructor.
@@ -80,7 +96,19 @@ class CallbackService {
      * @param OrderFactory $orderFactory
      * @param OrderRepositoryInterface $orderRepository
      */
-    public function __construct(CallbackMethod $callbackMethod, CallbackEventAdapter $eventAdapter, OrderFactory $orderFactory, OrderRepositoryInterface $orderRepository, GatewayConfig $gatewayConfig, InvoiceService $invoiceService, InvoiceRepositoryInterface $invoiceRepository, StoreCardService $storeCardService, CustomerFactory $customerFactory, StoreManagerInterface $storeManager) {
+    public function __construct(
+        CallbackMethod $callbackMethod,
+        CallbackEventAdapter $eventAdapter,
+        OrderFactory $orderFactory,
+        OrderRepositoryInterface $orderRepository,
+        GatewayConfig $gatewayConfig,
+        InvoiceService $invoiceService,
+        InvoiceRepositoryInterface $invoiceRepository,
+        StoreCardService $storeCardService,
+        CustomerFactory $customerFactory,
+        StoreManagerInterface $storeManager,
+        OrderSender $orderSender
+    ) {
         $this->callbackMethod   = $callbackMethod;
         $this->eventAdapter     = $eventAdapter;
         $this->orderFactory     = $orderFactory;
@@ -91,6 +119,7 @@ class CallbackService {
         $this->storeCardService    = $storeCardService;
         $this->customerFactory = $customerFactory;
         $this->storeManager = $storeManager;
+        $this->orderSender          = $orderSender;
     }
 
     /**
@@ -100,74 +129,78 @@ class CallbackService {
      * @throws LocalizedException
      */
     public function run() {
-        if( ! $this->gatewayResponse) {
+        if(!$this->gatewayResponse) {
             throw new DomainException('The response is empty.');
         }
 
+        // Set the gateway response
         $this->callbackMethod->setGatewayResponse($this->gatewayResponse);
 
+        // Perform tasks and prepare data
         $commandName    = $this->getCommandName();
         $amount         = $this->getAmount();
-
-        $this->callbackMethod->validate();
-
         $order      = $this->getAssociatedOrder();
         $payment    = $order->getPayment();
 
-        if($payment instanceof Payment) {
+        // Get override comments setting from config
+        $overrideComments = $this->gatewayConfig->overrideOrderComments();
+
+        // Process the payment
+        if ($payment instanceof Payment) {
+            // Prepare payment info
             $infoInstance = $payment->getMethodInstance()->getInfoInstance();
             $this->callbackMethod->setInfoInstance($infoInstance);
-
             $this->putGatewayResponseToHolder();
 
-            // Perform the required action on transaction
-            switch($commandName) {
-                case 'capture':
-                    $this->callbackMethod->capture($payment, $amount);
-                    break;
-                case 'refund':
-                    $this->callbackMethod->refund($payment, $amount);
-                    break;
-                case 'void':
-                    $this->callbackMethod->void($payment);
-                    break;
+            // Test the command name
+            if ($commandName == 'refund') {
+                $this->callbackMethod->refund($payment, $amount);
             }
-
+            else if ($commandName == 'void') {
+                $this->callbackMethod->void($payment);
+            }
             // Perform authorize complementary actions
-            if ($commandName == 'authorize') {
+            else if ($commandName == 'authorize') {
                 // Update order status
-                $order->setState(\Magento\Sales\Model\Order::STATE_NEW);
-                $order->setStatus(\Magento\Sales\Model\Order::STATE_PENDING_PAYMENT);
+                $order->setStatus($this->gatewayConfig->getOrderStatusAuthorized());
 
-                // Delete comments history
-                foreach ($order->getAllStatusHistory() as $orderComment) {
-                    $orderComment->delete();
-                } 
+                // Send the email
+                $this->orderSender->send($order);
 
                 // Set email sent
                 $order->setEmailSent(1);
 
-                // Create new comment
-                $newComment = 'Authorized amount of ' . ChargeAmountAdapter::getStoreAmountOfCurrency($this->gatewayResponse['response']['message']['value'], $this->gatewayResponse['response']['message']['currency']) . ' ' . $this->gatewayResponse['response']['message']['currency'] .' Transaction ID: ' . $this->gatewayResponse['response']['message']['id'];
+                // Comments override
+                if ($overrideComments) {
+                    // Delete comments history
+                    foreach ($order->getAllStatusHistory() as $orderComment) {
+                        $orderComment->delete();
+                    } 
 
-                // Add the new comment
-                $order->addStatusToHistory($order->getStatus(), $newComment, $notify = true);
+                    // Create new comment
+                    $newComment = 'Authorized amount of ' . ChargeAmountAdapter::getStoreAmountOfCurrency($this->gatewayResponse['response']['message']['value'], $this->gatewayResponse['response']['message']['currency']) . ' ' . $this->gatewayResponse['response']['message']['currency'] .' Transaction ID: ' . $this->gatewayResponse['response']['message']['id'];
 
+                    // Add the new comment
+                    $order->addStatusToHistory($order->getStatus(), $newComment, $notify = true);
+                }
             }
 
             // Perform capture complementary actions
-            if ($commandName == 'capture') {
+            else if ($commandName == 'capture') {
                 // Update order status
-                $order->setState(\Magento\Sales\Model\Order::STATE_NEW);
-                $order->setStatus($this->gatewayConfig->getNewOrderStatus());
+                $order->setStatus($this->gatewayConfig->getOrderStatusCaptured());
 
-                // Create new comment
-                $newComment = 'Captured amount of ' . ChargeAmountAdapter::getStoreAmountOfCurrency($this->gatewayResponse['response']['message']['value'], $this->gatewayResponse['response']['message']['currency']) . ' ' . $this->gatewayResponse['response']['message']['currency'] .' Transaction ID: ' . $this->gatewayResponse['response']['message']['id'];
+                // Comments override
+                if ($overrideComments) {
+                    // Create new comment
+                    $newComment = 'Captured amount of ' . ChargeAmountAdapter::getStoreAmountOfCurrency($this->gatewayResponse['response']['message']['value'], $this->gatewayResponse['response']['message']['currency']) . ' ' . $this->gatewayResponse['response']['message']['currency'] .' Transaction ID: ' . $this->gatewayResponse['response']['message']['id'];
 
-                // Add the new comment
-                $order->addStatusToHistory($order->getStatus(), $newComment, $notify = true);
+                    // Add the new comment
+                    $order->addStatusToHistory($order->getStatus(), $newComment, $notify = true);
+                }
             }
 
+            // Save the order
             $this->orderRepository->save($order);
         }
     }
